Created basic physical memory manager
Physical memory is now tracked by a bitmap. Memory can be allocated and freed, given the size of an object. The state of the memory map can also be printed. The memory is map by bootloader is also pushed onto the stack after boot, so that the memory manager can mark all the used/free areas as such.
Also added tty functionality for changing the color, either permanently of temporarily, until tty_reset_color is called.
And some minor refactoring.