Mapped kernel to higher half (3 GiB)
The boot.asm script now handles setting up paging. It first identity maps the first 4 MiB and then maps 4 MiB starting at 0xC0000000 (3 GiB) to the first 4 MiB.
After a long jump has been performed, the identity mapping is unloaded.
The rest of the kernel was adjusted to deal with the new addresses. The vmm initialization now consists only of setting the initially setup directory as the current one.

A bug was also fixed in the pmm, which would increase the used_blocks variable when running mark_used, even if the block was already marked as used.